Use PUB_SUB from AppModule
