+++ backend/src/app/graphql.module.ts
@@ -1,0 +1,7 @@
+import { GraphQLModule } from "@nestjs/graphql";
+
+export const graphQLModule = GraphQLModule.forRoot({
+  autoSchemaFile: `${process.cwd()}/src/schema.gql`,
+  installSubscriptionHandlers: true,
+  debug: true,
+});
