--- backend/src/constants.ts
@@ -1,3 +1,0 @@
-const PUB_SUB = Symbol("PUB_SUB");
-
-export { PUB_SUB };
